#44 [REFACTOR] 이미지 미리보기 URL 관리 로직을 useImagePreview hook으로 분리
